♻️ Rename atom to atoms
