+++ src/lib/components/atoms/SearchIcon.svelte
